third_party/blink/web_tests/svg/dom/points-parser-expected.txt

CONSOLE ERROR: Error: <polyline> attribute points: Expected number, "a".
CONSOLE ERROR: Error: <polyline> attribute points: Unexpected end of attribute. Expected number, "10".
CONSOLE ERROR: Error: <polyline> attribute points: Unexpected end of attribute. Expected number, "10,".
CONSOLE ERROR: Error: <polyline> attribute points: Expected number, "10,,".
CONSOLE ERROR: Error: <polyline> attribute points: Expected number, "10,,10".
CONSOLE ERROR: Error: <polyline> attribute points: Unexpected end of attribute. Expected number, "10,10,".
CONSOLE ERROR: Error: <polyline> attribute points: Expected number, "10,10a".
CONSOLE ERROR: Error: <polyline> attribute points: Expected number, ",10".
CONSOLE ERROR: Error: <polyline> attribute points: Expected number, ",10,".
CONSOLE ERROR: Error: <polyline> attribute points: Expected number, ",10,,".
CONSOLE ERROR: Error: <polyline> attribute points: Expected number, ",10,10".
This tests the parser for points attribute values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test string: ' '
PASS '' is ''


Test string: '10 10'
PASS '10 10' is '10 10'


Test string: '10,10'
PASS '10 10' is '10 10'


Test string: '10, 10'
PASS '10 10' is '10 10'


Test string: '10 10 20 20'
PASS '10 10 20 20' is '10 10 20 20'


Test string: '10, 10 20 20'
PASS '10 10 20 20' is '10 10 20 20'


Test string: '10, 10, 20 20'
PASS '10 10 20 20' is '10 10 20 20'


Test string: '10, 10, 20, 20'
PASS '10 10 20 20' is '10 10 20 20'


Test string: '10, 10 20, 20'
PASS '10 10 20 20' is '10 10 20 20'


Test string: '10 10, 20, 20'
PASS '10 10 20 20' is '10 10 20 20'


Test string: '10 10 20, 20'
PASS '10 10 20 20' is '10 10 20 20'


Test string: '10+10'
PASS '10 10' is '10 10'


Test string: '10-10'
PASS '10 -10' is '10 -10'


Test string: '10,+10'
PASS '10 10' is '10 10'


Test string: '10,-10'
PASS '10 -10' is '10 -10'


Test string: '            10,10'
PASS '10 10' is '10 10'


Test string: '10,10            '
PASS '10 10' is '10 10'


Test string: '10,            10'
PASS '10 10' is '10 10'


Test string: '10            ,10'
PASS '10 10' is '10 10'


Test string: '10     ,       10'
PASS '10 10' is '10 10'


Test string: '			10,10'
PASS '10 10' is '10 10'


Test string: '10,10			'
PASS '10 10' is '10 10'


Test string: '10,			10'
PASS '10 10' is '10 10'


Test string: '10			,10'
PASS '10 10' is '10 10'


Test string: '10		,	10'
PASS '10 10' is '10 10'


Strings that fail to parse, and hence parse as "".


Test string: 'a'
PASS '' is ''


Test string: '10'
PASS '' is ''


Test string: '10,'
PASS '' is ''


Test string: '10,,'
PASS '' is ''


Test string: '10,,10'
PASS '' is ''


Test string: '10,10,'
PASS '10 10' is '10 10'


Test string: '10,10a'
PASS '10 10' is '10 10'


Test string: ',10'
PASS '' is ''


Test string: ',10,'
PASS '' is ''


Test string: ',10,,'
PASS '' is ''


Test string: ',10,10'
PASS '' is ''


PASS successfullyParsed is true

TEST COMPLETE
